sdk/python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "marmot-sdk"
version = "0.1.0"
description = "Python SDK for Marmot — read the catalog and self-document automation, agents, and pipelines."
requires-python = ">=3.10"
license = { text = "Apache-2.0" }
authors = [{ name = "Marmot" }]
keywords = ["marmot", "catalog", "data", "lineage", "agents", "metadata"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "httpx>=0.27",
    "pyyaml>=6.0",
    "typing-extensions>=4.9; python_version < '3.12'",
]

[project.urls]
Homepage = "https://marmot.run"
Repository = "https://github.com/marmotdata/marmot"
Issues = "https://github.com/marmotdata/marmot/issues"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-httpx>=0.30",
    "ruff>=0.6",
    "openapi-python-client>=0.21",
]

[tool.hatch.build.targets.wheel]
packages = ["src/marmot"]

[tool.ruff]
line-length = 100
target-version = "py310"
src = ["src", "tests"]
extend-exclude = ["src/marmot/_gen"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM", "RUF"]
ignore = ["E501"]  # line length handled by formatter

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra --strict-markers"
